automaton/requirements.txt
aws-cdk-lib
constructs
jsii
networkx
